Refactor, remove manual entry option
