+++ Pipfile
@@ -7,7 +7,7 @@ name = "pypi"
 requests = "*"
 
 [dev-packages]
-black = "==20.8b1"
+black = "==21.10b0"
 flake8 = "*"
 mypy = "*"
