--- common.py
@@ -1,7 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class Update:
-    station: str
-    value: int
